+++ dev-packages/cli/src/info/info.ts
@@ -1,5 +1,5 @@
 import { CliContext, ContextUtils } from '@malagu/cli-common/lib/context/context-protocol';
-import { HookExecutor } from '@malagu/cli-common/lib/hook/hook-executor';
+import { HookExecutor, HookStage } from '@malagu/cli-common/lib/hook/hook-executor';
 import { LoggerUtil } from '@malagu/cli-common/lib/utils/logger-util';
 
 export interface InfoOptions {}
@@ -8,13 +8,14 @@ export default async (cliContext: CliContext, options: InfoOptions) => {
     try {
         cliContext.options = options;
         const ctx = await ContextUtils.createInfoContext(cliContext);
+        const hookExecutor = new HookExecutor();
+        await hookExecutor.executeInfoHooks(ctx, HookStage.before);
 
         LoggerUtil.printStage(ctx);
         LoggerUtil.printMode(ctx);
         LoggerUtil.printTargets(ctx);
         LoggerUtil.printComponents(ctx);
 
-        const hookExecutor = new HookExecutor();
         await hookExecutor.executeInfoHooks(ctx);
     } catch (error) {
         console.error(error);