+++ dev-packages/cli/templates/nest-app/src/main.ts
@@ -1,8 +1,4 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 
-async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-  await app.listen(9000);
-}
-bootstrap();
+export default NestFactory.create(AppModule);