+++ plugins/fc-plugin/src/hooks/props.ts
@@ -1,9 +1,10 @@
 import { PropsContext } from '@malagu/cli-common';
+import { BACKEND_TARGET } from '@malagu/cli-common/lib/constants';
 import { CloudUtils, DefaultProfileProvider } from '@malagu/cloud-plugin';
 
-export default async (context: PropsContext) => {
-    const { props, cfg } = context;
-    if (props.mode && props.mode.includes('remote')) {
+export async function before(context: PropsContext) {
+    const { props, cfg, target } = context;
+    if (target === BACKEND_TARGET && props.mode && props.mode.includes('remote')) {
         context.spinner?.stop();
         const cloudConfig = CloudUtils.getConfiguration(cfg);
         const profileProvider = new DefaultProfileProvider();